requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
Flask-Bcrypt==1.0.1
PyMySQL==1.1.1
email-validator==2.2.0
python-dotenv==1.0.1
